refactor(profiles): split monitor matching out of ProfileManager
The case-insensitive freeName expansion pushed the file past
the 350-line ceiling; the monitor-matching queries (match,
defaultProfile, duplicateDefaultCounts) are a cohesive pure
block and move to a sibling extension.
